update git ignore and rename litecoin-api to suit convention
