+++ api/litecoin-api.js
